# backend/openapi.yaml
openapi: 3.0.3
info:
  title: WuXuxian TTRPG API
  description: |
    RESTful API for the WuXuxian TTRPG webapp.
    
    Provides endpoints for:
    - Character management (PCs, NPCs, bosses)
    - Technique/ability management
    - Boss template management
    - Combat simulation configuration and execution
    
    The API follows RESTful conventions and uses JSON for request/response payloads.
  version: 0.0.1
  contact:
    name: WuXuxian TTRPG Team
  license:
    name: MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://api.wuxuxian.example.com/api/v1
    description: Production server (example)
...
paths:
...
  /techniques:
    get:
      summary: List all techniques
      description: Retrieve a list of all techniques with optional filters
      operationId: listTechniques
      tags:
        - techniques
      parameters:
        - name: technique_type
          in: query
          schema:
            type: string
            enum: [Basic, Standard, Major, Spike]
        - name: is_quick_action
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of techniques
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Technique'
    
    post:
      summary: Create a new technique
      description: Create a new technique/ability
      operationId: createTechnique
      tags:
        - techniques
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TechniqueCreate'
      responses:
        '201':
          description: Technique created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Technique'
  /techniques/{technique_id}:
    get:
      summary: Get a technique by ID
      operationId: getTechnique
      tags:
        - techniques
      parameters:
        - name: technique_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Technique details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Technique'
        '404':
          description: Technique not found
    
    patch:
      summary: Update a technique
      operationId: updateTechnique
      tags:
        - techniques
      parameters:
        - name: technique_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TechniqueUpdate'
      responses:
        '200':
          description: Technique updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Technique'
        '404':
          description: Technique not found
    
    delete:
      summary: Delete a technique
      operationId: deleteTechnique
      tags:
        - techniques
      parameters:
        - name: technique_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Technique deleted successfully
        '404':
          description: Technique not found
...
components:
  schemas:
...
    Technique:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        technique_type:
          type: string
          enum: [Basic, Standard, Major, Spike]
          nullable: true
        base_damage:
          type: integer
        ae_cost:
          type: integer
        self_strain:
          type: integer
        damage_routing:
          type: string
          enum: [THP, Guard, Strain]
        boss_strain_on_hit:
          type: integer
        dr_debuff:
          type: number
          format: float
        is_quick_action:
          type: integer
    TechniqueCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        technique_type:
          type: string
          enum: [Basic, Standard, Major, Spike]
          nullable: true
        base_damage:
          type: integer
        ae_cost:
          type: integer
        self_strain:
          type: integer
        damage_routing:
          type: string
          enum: [THP, Guard, Strain]
        boss_strain_on_hit:
          type: integer
        dr_debuff:
          type: number
          format: float
        is_quick_action:
          type: integer
    TechniqueUpdate:
      type: object
      properties:
        name:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        technique_type:
          type: string
          enum: [Basic, Standard, Major, Spike]
          nullable: true
        base_damage:
          type: integer
          nullable: true
        ae_cost:
          type: integer
          nullable: true
        self_strain:
          type: integer
          nullable: true
        damage_routing:
          type: string
          enum: [THP, Guard, Strain]
          nullable: true
        boss_strain_on_hit:
          type: integer
          nullable: true
        dr_debuff:
          type: number
          format: float
          nullable: true
        is_quick_action:
          type: integer
          nullable: true
...